do not use autogenerated libs variable
The autogenerated one contaisn every opencv library, including the optional ones.
Which in turn can cause issues on systems where the optional ones are not installed.

So we pick only the ones we actually need, which we used to do originally as well.
However that also included opencv_features2d, presumably left by someone else. That, as defined in OpenCVModules.cmake did depend on opencv_highgui, which depends on gtk3.
Resulting in undefined reference errors on build.(on some systems).